✨ feat(route.ts): add PUT and DELETE functions to handle updating and deleting posts
The PUT function allows updating a post by its ID. It receives the request and the ID of the post to be updated, and updates the title and content of the post in the database.

The DELETE function allows deleting a post by its ID. It receives the request and the ID of the post to be deleted, and deletes the post from the database.